% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allinculdedfilters}
\alias{allinculdedfilters}
\title{All identified filters.}
\format{
67 observations and 3 variables
\describe{
\item{code}{Code identifying the filter.}
\item{filter}{The filter statement.}
\item{source}{The source of the filter, refered to the ID in sources.}
...
}
}
\usage{
allinculdedfilters
}
\description{
All identified filters.
}
\keyword{datasets}
