% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{originalfilters}
\alias{originalfilters}
\title{Original selected filters used in the survey.}
\format{
26 observations and 3 variables
\describe{
\item{code}{Code identifying the filter.}
\item{filter}{The filter statement.}
\item{source}{The source of the filter, refered to the ID in sources.}
...
}
}
\usage{
originalfilters
}
\description{
Original selected filters used in the survey.
}
\keyword{datasets}
