% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results}
\alias{results}
\title{Final results of the Delphi trial.}
\format{
67 observations and 8 variables
\describe{
\item{code}{Referring to the ID for the individual filter.}
\item{filter}{Filter statement.}
\item{median_score}{The final median score of the filter, in the last round it was applied.}
\item{pvalue}{p-value in comparing significant change between rounds.}
\item{source}{Initial source of the filter as specified in sources.}
\item{round}{Final round when the filter reached it's end state.}
\item{site}{ID of the study-site that the filter originated from.}
\item{modified}{Indicated if the filter was modified Y/N/NA.}
...
}
}
\usage{
results
}
\description{
Final results of the Delphi trial.
}
\keyword{datasets}
