% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sources}
\alias{sources}
\title{Sources for audit filters.}
\format{
5 observations and 3 variables
\describe{
\item{id}{Assigned source id.}
\item{description}{Description of the source for the paper, example, WHO.}
\item{reference}{Reference to the publication defining the filters.}
...
}
}
\usage{
sources
}
\description{
Sources for audit filters.
}
\keyword{datasets}
